pyproject.toml
[project]
name = "ai-finance-advisor"
version = "0.1.0"
readme = "README.md"
requires-python = ">=3.12,<4.0"
dependencies = [
    "langchain (>=0.3.19,<0.4.0)",
    "langchain-community (>=0.3.18,<0.4.0)",
    "langchain-openai (>=0.3.7,<0.4.0)",
    "supabase (>=2.13.0,<3.0.0)",
    "python-dotenv (>=1.0.1,<2.0.0)",
    "faster-whisper (>=1.1.1,<2.0.0)",
    "streamlit (>=1.43.0,<2.0.0)",
    "gradio (>=5.20.1,<6.0.0)"
]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
package-mode = false
